+++ modules/contentbox/migrations/2022_05_27_133613_CONTENTBOX-1437-Drop-cacheLayout-Column.cfc
@@ -1,6 +1,5 @@
 /**
- * This migration will remove the 'SSLOnly' column from the 'cb_page' table because this bit was not nullable,
- * had a default value and we won't use it anymore, So we need to remove it, else all page editing will fail.
+ * This migration will remove the 'cacheLayout' column from the 'cb_content' as it is no longer in use
  *
  * @see     CONTENTBOX-1437 - https://ortussolutions.atlassian.net/browse/CONTENTBOX-1437
  * @version 5.3.0
@@ -8,14 +7,19 @@
  */
 component {
 
+	// DI
+	property name="migrationService" inject="MigrationService@cfmigrations";
+
+	// Include Utils
+	include template="./_MigrationUtils.cfm";
+
 	function up( schema, qb ){
-		try {
-			// Remove the sslonly column from the `cb_page` table
+		if ( schema.hasColumn( "cb_content", "cacheLayout" ) ) {
 			schema.alter( "cb_content", ( table ) => {
 				table.dropColumn( "cacheLayout" );
 			} );
-		} catch ( any e ) {
-			// If this query errors, then the schema was created by a version greater than 5.0
+		} else {
+			systemOutput( "- skipping 'cacheLayout' removal, cb_content doesn't have the column", true );
 		}
 	}
